backend/requirements.txt
astroid==3.0.3
autopep8==2.0.4
bcrypt==4.0.1
black==23.7.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.4
dill==0.3.9
distlib==0.3.9
dnspython==2.7.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.2
packaging==24.1
pandas==2.1.0
pathspec==0.12.1
pbr==6.1.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==4.21.12
pycodestyle==2.12.1
pylint==3.0.1
pymongo==4.10.1
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.32.3
scikit-surprise==1.1.4
scipy==1.13.1
six==1.16.0
stevedore==5.3.0
surprise==0.1
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
Werkzeug==3.0.4
zipp==3.20.2
